chore: convert request-system-cas to use PlatformSpecific
